Event bus with promises, profile and settings works
