+++ src/ui/templates/about/about.pug
@@ -9,7 +9,7 @@
 				a(href='https://t.me/BarulevAlex',target='_blank').contakt-link
 					img(src=require('../../images/tg.svg')).contact-link__icon
 				a(href='https://github.com/HustonMmmavr',target='_blank').contakt-link
-					img(srcrequire('../../images/git.svg')).contact-link__icon
+					img(src=require('../../images/git.svg')).contact-link__icon
 		.team-member
 			img(src=require('../../images/avatars/zaytsev.jpg')).team-member__img
 			.team-member__description
