+++ src/ui/templates/form/form.pug
@@ -15,8 +15,8 @@ if back
 			.input__holder
 				if input.type == 'file'
 					label.file-upload
-						span.file-upload__button.btn Choose photo
-						mark.file-upload__path No photo to upload
+						span.file-upload__button.btn CHOOSE FILE
+						mark.file-upload__path No file to upload
 						input(type=input.type,name=input.name).form__input
 						span(class='error-'+input.name).error-message
 				else
