+++ src/ui/templates/header/header.pug
@@ -1,6 +1,6 @@
 header.header
 	.header__container.header__container_left
-		.header__logo
+		.header__logo(hidden=true)
 			a(href='/').header__logo__link
 				| C
 				img(src=require('../../images/smile.svg')).header__logo__smile
